requirements.txt
atlassian-python-api==3.11.0